MYR-383 Split the wire types and the accept event projection out of two files
Both were within a handful of lines of the 300-line cap and both needed room
for the window gate. Pure moves: ride_request_wire_types.go takes the JSON
shapes off ride_request_types.go, ride_accepted_event.go takes the dispatch-seam
projection off ride_request_owner_handler.go. No behaviour moves with them.
